Minor tweaks to README.md.
Changed the environment variable JAGUAR_PATH to JAG_PATH.
This doesn't seem to be used anywhere, only in the README,
so I don't think it needs updating anywhere else.  The change
makes us consistent in using JAG_ as a prefix for environment
variables.

Put in a pointer from the top to the bottom of the README for those
that would rather build their own open source projects, rather than
download binaries.