Set server timeout to 3s, roll to new HTTP pkg.
When we load the mini web page in a browser, the browser
will tend to leave the connection open for a long time.
With this change we quickly time out the connection so that
the jag command line is not blocked (we run with only one
connection at a time on the devices).

This also rolls to the new HTTP pkg version 2.2.0.