Fix device config storage with ints
We wouldn't recognize the stored 8 and return the default 4 all the
time.

The json-decoder always uses a 64-bit float which is what the function
was written for. However, the config file stores the value as integer,
    and we would thus ignore that value.

Fixes `jag simulate` without host-name, since that one typically has 8
bytes for a word.